+++ NEWS.md
@@ -41,6 +41,7 @@ __New Recipes Preprocessing Steps__:
 __Improvements__:
 
 * `plot_acf_diagnostics()`: The `.lags` argument now handles time-based phrases (e.g. `.lags = "1 month"`).
+* `time_series_cv()`: Implements time-based phrases (e.g. `initial = "5 years"` and `assess = "1 year"`)
 
 __Bug Fixes__:
 
@@ -64,6 +65,7 @@ These should not be of major impact since the 1.0.0 version was just released.
 * Renamed `roll_apply_vec()` to `slidify_vec()` for consistency with `slidify()` & relationship to `slider` R package
 * Renamed `step_roll_apply` to `step_slidify()` for consistency with `slidify()` & relationship to `slider` R package
 * Renamed `tk_augment_roll_apply` to `tk_augment_slidify()` for consistency with `slidify()` & relationship to `slider` R package
+
 
 # timetk 1.0.0